refactor(krt): one round helper serves both test modules
The `round` fixture stood twice: once in the test module of `stats.rs` and
once in the test module of `main.rs`. The two bodies were the same twenty
lines, and they differed only in the three constants they read for the run
id, the moment, and the duration. The `address` fixture stood six times,
byte for byte the same in every test module of the crate.

Both now live in one `#[cfg(test)] mod testing` at `src/krt/src/testing.rs`,
which `main.rs` declares beside the other modules. The shared module holds
the run id, the moment, the round duration, and the ICMP name as private
constants, so the fixture keeps one narrow entrance and its callers name
only the TTL range and the hops.

The two sets of constants held different values — `RUN`/`MOMENT` at
2026-08-18T12:00:00.000Z, and `AGGREGATE_RUN`/`AGGREGATE_MOMENT` at
2026-08-20T00:00:00.000Z. No test reads either value: each constant reached
its own copy of `round` and nothing else, so the shared module takes the
`stats.rs` pair, which matches the fixtures of the other modules of the
crate. `AGGREGATE_RUN`, `AGGREGATE_MOMENT`, `AGGREGATE_DURATION`, both
copies of `TIME_EXCEEDED`, and the `RUN`, `MOMENT`, and `ROUND_DURATION` of
`stats.rs` are gone, along with the imports that only they reached.

The note about why the resolution tests name a literal address and never a
host name moved onto `socket`, which is the helper of those tests.

No test accompanies this change. It moves test code and changes no behavior
of the tool, so the existing suite is the check: 257 unit tests, 12 in
`tests/cli.rs`, and 12 in `tests/replay.rs`, the same counts as before, all
passing and none of them changed in meaning.
